Makes client_max_body_size configurable
The default config was at 4G and offered no configuration. The default
nginx config is 1m. 4G is a huge limit and unsuitable for most installs.

I've set it to 10m as the default which is more than nginx's default,
but still reasonable in comparison to the 4G limit. If you, for whatever
reason, need a different value, simply pass it to nginx::site via the
client_max_body_size parameter.